requirements.txt
pytest
openpyxl
allure-pytest
loguru
requests